add batch tests
